backend/requirements.txt
# Core
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# LLM - LiteLLM (multi-provider gateway)
litellm>=1.50.0

# Object Storage (S3-compatible)
minio>=7.2.0

# Database
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Data Sources
arxiv>=2.1.0
requests>=2.31.0

# Utilities
tenacity>=8.2.0
cachetools>=5.3.0
python-dateutil>=2.9.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.30.0

# Output
markdown>=3.6
fpdf2>=2.7.9
python-docx>=1.1.2

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0

# Code Quality
black>=24.0.0
flake8>=7.0.0
mypy>=1.10.0
